Changed word query to use crime data
Moved all the Docker/Gradle stuff into a `setup` directory for now.

Did not change the other two retrievers, as this PR was already pretty big. Will do those next.
