--- rag-langchain-python/build.gradle
@@ -1,4 +1,0 @@
-plugins {
-    id "net.saliman.properties" version "1.5.2"
-    id "com.marklogic.ml-gradle" version "4.6.0"
-}
