--- rag-langchain-python/docker-compose.yml
@@ -1,17 +1,0 @@
-version: '3.8'
-name: marklogic_python_example_langchain
-
-services:
-
-  marklogic:
-    image: "marklogicdb/marklogic-db:11.1.0-centos-1.1.0"
-    platform: linux/amd64
-    environment:
-      - MARKLOGIC_INIT=true
-      - MARKLOGIC_ADMIN_USERNAME=admin
-      - MARKLOGIC_ADMIN_PASSWORD=admin
-    volumes:
-      - ./docker/marklogic/logs:/var/opt/MarkLogic/Logs
-    ports:
-      - "8000-8003:8000-8003"
-
