--- rag-langchain-python/gradle.properties
@@ -1,4 +1,0 @@
-mlAppName=langchain-test
-mlRestPort=8003
-mlUsername=admin
-mlPassword=admin
