--- rag-langchain-python/src/main/ml-config/databases/content-database.json
@@ -1,4 +1,0 @@
-{
-  "database-name": "%%DATABASE%%",
-  "schema-database": "%%SCHEMAS_DATABASE%%"
-}
