--- rag-langchain-python/src/main/ml-config/databases/schemas-database.json
@@ -1,3 +1,0 @@
-{
-  "database-name": "%%SCHEMAS_DATABASE%%"
-}
