--- rag-langchain-python/src/main/ml-config/security/roles/langchain-eval-role.json
@@ -1,20 +1,0 @@
-{
-  "role-name": "langchain-eval-role",
-  "privilege": [
-    {
-      "privilege-name": "xdmp:eval",
-      "action": "http://marklogic.com/xdmp/privileges/xdmp-eval",
-      "kind": "execute"
-    },
-    {
-      "privilege-name": "xdmp:eval-in",
-      "action": "http://marklogic.com/xdmp/privileges/xdmp-eval-in",
-      "kind": "execute"
-    },
-    {
-      "privilege-name": "xdbc:eval",
-      "action": "http://marklogic.com/xdmp/privileges/xdbc-eval",
-      "kind": "execute"
-    }
-  ]
-}
