--- rag-langchain-python/src/main/ml-config/security/users/langchain-user.json
@@ -1,10 +1,0 @@
-{
-    "user-name": "langchain-user",
-    "password": "password",
-    "role": [
-        "rest-reader",
-        "rest-writer",
-        "qconsole-user",
-        "langchain-eval-role"
-    ]
-}
