+++ setup/docker-compose.yml
@@ -1,0 +1,17 @@
+version: '3.8'
+name: marklogic_ai_examples
+
+services:
+
+  marklogic:
+    image: "marklogicdb/marklogic-db:latest-11"
+    platform: linux/amd64
+    environment:
+      - MARKLOGIC_INIT=true
+      - MARKLOGIC_ADMIN_USERNAME=admin
+      - MARKLOGIC_ADMIN_PASSWORD=admin
+    volumes:
+      - ./docker/marklogic/logs:/var/opt/MarkLogic/Logs
+    ports:
+      - "8000-8003:8000-8003"
+
