+++ setup/gradle.properties
@@ -1,0 +1,5 @@
+mlAppName=langchain-test
+mlHost=localhost
+mlRestPort=8003
+mlUsername=admin
+mlPassword=admin
