+++ setup/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Tue Mar 22 14:27:38 EDT 2016
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.4-bin.zip
