+++ setup/src/main/ml-config/databases/content-database.json
@@ -1,0 +1,4 @@
+{
+  "database-name": "%%DATABASE%%",
+  "schema-database": "%%SCHEMAS_DATABASE%%"
+}
