+++ setup/src/main/ml-config/databases/schemas-database.json
@@ -1,0 +1,3 @@
+{
+  "database-name": "%%SCHEMAS_DATABASE%%"
+}
