+++ setup/src/main/ml-config/security/users/langchain-user.json
@@ -1,0 +1,10 @@
+{
+    "user-name": "langchain-user",
+    "password": "password",
+    "role": [
+        "rest-reader",
+        "rest-writer",
+        "qconsole-user",
+        "langchain-eval-role"
+    ]
+}
